+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## [0.4.0](https://github.com/runpod/tetra-rp/compare/v0.3.0...v0.4.0) (2025-06-26)
+
+
+### Features
+
+* CPU Endpoints and Live Serverless ([c335fd1](https://github.com/runpod/tetra-rp/commit/c335fd11cbc65c4a18edabe007982506bee5e6c3))
+
+
+### Bug Fixes
+
+* CpuInstanceType values ([9cf958b](https://github.com/runpod/tetra-rp/commit/9cf958b381d56b4aa5d42c0393500818d6de3c34))
+
 ## [0.3.0](https://github.com/runpod/tetra-rp/compare/v0.2.1...v0.3.0) (2025-06-23)
 
 
@@ -29,7 +41,7 @@
 * is_deployed() to check and redeploy in case endpoint is inaccessible ([4aa0932](https://github.com/runpod/tetra-rp/commit/4aa0932c9f99b08a0d234d4d843dd6339cbc7dd1))
 * Prefer A40 for GPU list when "any" is set ([6a60418](https://github.com/runpod/tetra-rp/commit/6a60418452a39501b3142096e9ab8afa01628953))
 * preparing json.dump of complex objects with normalize_for_json ([48cedd6](https://github.com/runpod/tetra-rp/commit/48cedd64e9140d261f2b3327a3611715d2bf3e38))
-* Pydantic modeling of RunPod resources ([87067bb](https://github.com/runpod/tetra-rp/commit/87067bbae31d7de618d0edd1d78289fef1c0f6d1))
+* Pydantic modeling of Runpod resources ([87067bb](https://github.com/runpod/tetra-rp/commit/87067bbae31d7de618d0edd1d78289fef1c0f6d1))
 * remote stub for ServerlessEndpoint calls ([739aa93](https://github.com/runpod/tetra-rp/commit/739aa93775e296fbb99aa9afccc07b12ae60ea50))
 * resilient redeployments ([fd2e78d](https://github.com/runpod/tetra-rp/commit/fd2e78d42531f83cc3b66d197fb2195244d65336))
 * resilient serverless calls with retries and auto-cancellations when unhealthy/throttled ([9f53e8c](https://github.com/runpod/tetra-rp/commit/9f53e8cb36871ff58b491cd8bf1851bf3d1029cc))